Use uint8 for segmentation labels
I don't think we'll ever have more than 256 classes and this will cut
memory and disk usage by 4x.

+++ rastervision/data/label/semantic_segmentation_labels.py
@@ -43,7 +43,7 @@ class SemanticSegmentationLabels(Labels):
         """
         arr = self.to_array()
         mask = rasterize(
-            [(p, 1) for p in aoi_polygons], out_shape=arr.shape, fill=0)
+            [(p, 1) for p in aoi_polygons], out_shape=arr.shape, fill=0, dtype=np.uint8)
         arr = arr * mask
         return SemanticSegmentationLabels.from_array(arr)
